+++ utils/agent_state.py
@@ -253,7 +253,7 @@ class AgentState:
         Update local objective list with the global coordinates of the local objective of each agent
         :param goal_pos_list: list of agents goal position, np.ndarray, shape=(num_agents, 2)
         :param agent_pos_list: list of agents current position, np.ndarray, shape=(num_agents, 2)
-        :return: torch Int Tensor of the input configuration
+        :return: torch Float Tensor of the input configuration
                  input_tensor.shape = (num_agents, 3 (channels), FOV+2*border, FOV+2*border)
         """
         # get map with agents positions, padded in the outside
@@ -274,7 +274,7 @@ class AgentState:
 
         # transform input state (list of ndarray) into a ndarray,
         # since creating a tensor from a list of ndarray is extremely slow
-        input_tensor = torch.from_numpy(np.array(input_state)).int()
+        input_tensor = torch.from_numpy(np.array(input_state)).float()
 
         return input_tensor
 
@@ -287,7 +287,7 @@ class AgentState:
         :param agent_pos_schedule: schedule (sequence) of agents positions
                                    np.ndarray, shape = (makespan, num_agents, 2)
         :param makespan: length of the agents schedule
-        :return: torch Int Tensor of the input configuration
+        :return: torch Float Tensor of the input configuration
                  input state = makespan x num_agents x input state of agent
         """
         input_step_list = []
@@ -314,6 +314,6 @@ class AgentState:
 
         # transform input state (list of ndarray) into a ndarray,
         # since creating a tensor from a list of ndarray is extremely slow
-        input_tensor = torch.from_numpy(np.array(input_step_list)).int()
+        input_tensor = torch.from_numpy(np.array(input_step_list)).float()
 
         return input_tensor
